lark-skygpt-v3-patched/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
gunicorn==22.0.0
python-dotenv==1.0.1
httpx==0.27.0
openai==1.30.1
apscheduler==3.10.4
pydantic-settings==2.2.1

# Document parsing
pypdf==4.2.0
python-docx==1.1.2
openpyxl==3.1.2

# Database & cache
asyncpg==0.29.0
SQLAlchemy[asyncio]==2.0.30
redis[asyncio]==5.0.4
